refactor(scanning): extract six helper clusters from the scan hub
Structural move: scanning/mod.rs (3003 lines) -> 1477, by extracting the
self-contained helper clusters the worker calls into sibling modules —
- payload_families.rs  payload catalog/pruning/interleave, DOM payloads,
                       inject-type labels
- waf_strategy.rs      WAF strategy selection + payload expansion + meta-CSP
- request_render.rs    build_request_text (the PoC request renderer)
- ast_dom_phase.rs     run_ast_dom_analysis + accumulate_findings +
                       external-JS fetch
- param_jobs.rs        generate_param_jobs + scannability + request estimate
- progress.rs          progress-bar UI + req/sec telemetry + WAF-block stats

The orchestration core (ScanWorkerCtx + its phases, FoundParams, ParamScanState,
run_scanning, ScanRunHandles, result-collapsing) stays in mod.rs — it is one
cohesive concern and run_scanning builds ScanWorkerCtx by struct literal, so
extracting it would force exposing all its private state. Extracted items are
pub(crate) and re-exported (`pub(crate) use <cluster>::*`) so every
`crate::scanning::X` path resolves unchanged; helper files share imports via
`use super::*`; tests.rs untouched.

Also removes the dead `csp_requires_trusted_types` (+ its 3 tests): verified
never called in production (the surfacing warning is what exposed it), a Batch-2
cleanup pulled forward since it cannot affect scan behavior.

1065 scanning tests pass, clippy + rustfmt clean.
