+++ src/scanning/mod.rs
@@ -68,6 +68,21 @@ pub mod vuln_libs;
 pub mod xss_blind;
 pub mod xss_common;
 
+// Scan-hub helper modules (extracted for parallel work / navigability).
+mod ast_dom_phase;
+mod param_jobs;
+mod payload_families;
+mod progress;
+mod request_render;
+mod waf_strategy;
+
+pub(crate) use ast_dom_phase::*;
+pub(crate) use param_jobs::*;
+pub(crate) use payload_families::*;
+pub(crate) use progress::*;
+pub(crate) use request_render::*;
+pub(crate) use waf_strategy::*;
+
 use crate::cmd::scan::ScanArgs;
 use crate::parameter_analysis::Param;
 use crate::scanning::check_dom_verification::check_dom_verification_with_client_outcome;
@@ -80,68 +95,6 @@ use std::sync::Arc;
 use std::sync::atomic::{AtomicU32, AtomicUsize, Ordering};
 use std::time::Duration;
 use tokio::sync::{Mutex, RwLock, Semaphore};
-
-/// Maximum number of WAF mutation variants generated per base payload.
-/// Prevents payload explosion when WAF bypass mutations are applied. This cap
-/// only takes effect when a WAF is detected and bypass is on (the only path
-/// that calls `expand_waf_payloads`), so it scales effort exactly on the scans
-/// that want more bypass attempts — not the common no-WAF path. Sized to give
-/// the attribute-decode-layer mutations (KeywordEntityEncode / MultiSlash /
-/// SchemeBreak) a slot alongside the proven structural ones for payloads where
-/// many mutations shape-match.
-const MAX_WAF_MUTATION_VARIANTS_PER_PAYLOAD: usize = 4;
-
-/// Maximum number of distinct same-origin external JS files fetched per target page fetch.
-const MAX_EXTERNAL_JS_FILES: usize = 16;
-/// Maximum bytes read from a single external JS file (matches analyzer limit).
-const MAX_EXTERNAL_JS_BYTES: usize = 512 * 1024;
-
-/// Build a `with_key("req_per_sec", …)` tracker for an indicatif progress bar.
-///
-/// `start` is the value of `crate::REQUEST_COUNT` captured at bar creation;
-/// the tracker renders `(REQUEST_COUNT - start) / pb.elapsed()` as
-/// `XXXX.X req/s` with a fixed-width field so the bar's trailing columns
-/// don't jitter as the rate changes magnitude.
-///
-/// Caveats baked into the displayed value:
-///   - `REQUEST_COUNT` is process-global, so when several targets in the
-///     same host group scan concurrently the per-target bar reflects the
-///     group's combined HTTP rate (which matches the overall bar). This is
-///     intentionally a "combined" view — a strictly per-target counter
-///     would need plumbing through every HTTP call site.
-///   - `{eta}` next to this field is still computed by indicatif from
-///     `pos/len` rate, not request rate. In practice ETA still reads
-///     sensibly because the bar finishes the moment the inner loop exits.
-pub(crate) fn req_per_sec_tracker(
-    start: u64,
-) -> impl Fn(&indicatif::ProgressState, &mut dyn std::fmt::Write) + Send + Sync + Clone + 'static {
-    move |state, w| {
-        let delta = crate::REQUEST_COUNT
-            .load(Ordering::Relaxed)
-            .saturating_sub(start);
-        let _ = write!(
-            w,
-            "{}",
-            format_req_per_sec(delta, state.elapsed().as_secs_f64())
-        );
-    }
-}
-
-/// Format `delta` requests over `elapsed_secs` as the right-aligned
-/// `XXXX.X req/s` field rendered by [`req_per_sec_tracker`]. Extracted as
-/// a pure helper so the rate / formatting contract can be tested without
-/// constructing an `indicatif::ProgressState`.
-///
-/// `elapsed_secs <= 0.0` yields `0.0 req/s` (avoids div-by-zero on the
-/// first tick before the bar has accumulated any duration).
-pub(crate) fn format_req_per_sec(delta: u64, elapsed_secs: f64) -> String {
-    let rate = if elapsed_secs > 0.0 {
-        delta as f64 / elapsed_secs
-    } else {
-        0.0
-    };
-    format!("{:>7.1} req/s", rate)
-}
 
 /// A per-parameter work unit for the scan loop: the parameter, its reflection
 /// payloads (checked in Stage 5), and its DOM payloads (verified in Stage 6).
@@ -193,1485 +146,6 @@ fn found_param_key(param: &Param) -> String {
         param.name,
         param.effective_wire_name()
     )
-}
-
-/// Label written to `Result.inject_type` for findings produced by the scan
-/// loop. Findings under `--sxss` are prefixed so JSON / markdown / plain
-/// reports distinguish stored from reflected results — downstream tooling
-/// parses this field, so the contract is pinned by `tests::test_inject_type_label_for_sxss`.
-fn inject_type_label_for(sxss: bool) -> &'static str {
-    if sxss { "sxss-inHTML" } else { "inHTML" }
-}
-
-/// True when the payload that produced the finding looks like a
-/// client-side template interpolation. `{{` / `}}` is sufficient on its
-/// own — Mustache, Handlebars, AngularJS, Vue and Ember all share that
-/// delimiter. Used to refine `inject_type` to `*-CSTI` so plain / JSON
-/// output makes the framework-injection finding distinguishable from a
-/// generic HTML reflection.
-fn is_template_shaped_payload(payload: &str) -> bool {
-    payload.contains("{{") && payload.contains("}}")
-}
-
-/// Map a framework innerHTML-sink directive name (recorded on
-/// `Param.framework_sink` during discovery) to the short suffix used in
-/// `inject_type`. Anything unrecognised falls back to a generic
-/// `-FrameworkSink` so the user still sees the class of finding even if
-/// dalfox grows support for a new directive name later.
-fn framework_sink_suffix(sink: &str) -> &'static str {
-    match sink {
-        "v-html" => "-VHtml",
-        "data-bind" => "-DataBind",
-        "ng-bind-html" => "-NgBindHtml",
-        "dangerouslySetInnerHTML" => "-DangerouslySetInnerHTML",
-        _ => "-FrameworkSink",
-    }
-}
-
-/// Refine the base `inject_type` label. Order of precedence:
-///   1. `-VHtml` / `-DataBind` / `-NgBindHtml` from a discovered
-///      framework innerHTML sink (highest signal — entity-encoded
-///      reflections in these attributes still execute).
-///   2. `-CSTI` for client-side template payloads (`{{ … }}`).
-///   3. base label only.
-///
-/// Mirrors the SXSS prefixing convention (`sxss-inHTML-VHtml`) so
-/// downstream parsers don't have to special-case ordering.
-#[cfg(test)]
-fn inject_type_for_payload(sxss: bool, payload: &str) -> String {
-    inject_type_for_payload_with_sink(sxss, payload, None)
-}
-
-fn inject_type_for_payload_with_sink(
-    sxss: bool,
-    payload: &str,
-    framework_sink: Option<&str>,
-) -> String {
-    let base = inject_type_label_for(sxss);
-    if let Some(sink) = framework_sink {
-        return format!("{}{}", base, framework_sink_suffix(sink));
-    }
-    if is_template_shaped_payload(payload) {
-        format!("{}-CSTI", base)
-    } else {
-        base.to_string()
-    }
-}
-
-fn reflection_kind_note(kind: crate::scanning::check_reflection::ReflectionKind) -> &'static str {
-    match kind {
-        crate::scanning::check_reflection::ReflectionKind::Raw => "reflected",
-        crate::scanning::check_reflection::ReflectionKind::HtmlEntityDecoded => {
-            "reflected after HTML-entity decoding"
-        }
-        crate::scanning::check_reflection::ReflectionKind::UrlDecoded => {
-            "reflected after URL/form decoding"
-        }
-        crate::scanning::check_reflection::ReflectionKind::HtmlThenUrlDecoded => {
-            "reflected after HTML-entity and URL/form decoding"
-        }
-    }
-}
-
-/// Drop payloads whose raw bytes carry an HTML-structural character the
-/// server has been observed to filter (recorded in
-/// `Param.invalid_specials` by Stage 3 active probing). Raw `<`/`>` cannot
-/// pass through a server-side blocklist that strips those bytes after
-/// decoding, so the corresponding payload has no chance to reflect — every
-/// HTTP request spent on it is wasted. The encoded variants of the same
-/// payload (`%3Csvg%3E`, `&lt;svg&gt;`, multi-URL-encoded forms) carry no
-/// raw `<`/`>` themselves and survive this pass, preserving the bypass
-/// surface for naive filters that decode only once.
-///
-/// Conservative on quotes: attribute-breakout payloads intentionally lead
-/// with the same delimiter character the surrounding HTML attribute uses,
-/// and that delimiter must already be a "valid" special (the server
-/// emitted it), so pruning on `"`/`'` would mistakenly drop the very
-/// payloads that exploit attribute injection.
-fn prune_blocked_raw_angles(payloads: Vec<String>, invalid_specials: &[char]) -> Vec<String> {
-    let block_lt = invalid_specials.contains(&'<');
-    let block_gt = invalid_specials.contains(&'>');
-    if !block_lt && !block_gt {
-        return payloads;
-    }
-    payloads
-        .into_iter()
-        .filter(|p| {
-            // Leading-window ("positional") filter bypass payloads carry raw
-            // `<`/`>` on purpose: their premise is that the block is positional,
-            // so the raw angle passes once the leading pad pushes it past the
-            // filtered window. Keep them despite the blocked-angle classification
-            // (they are FP-safe — a non-positional filter still encodes the `<`
-            // and nothing verifies). See `synthesis::positional_pad_payloads`.
-            crate::payload::synthesis::is_positional_pad_bypass(p)
-                || !((block_lt && p.contains('<')) || (block_gt && p.contains('>')))
-        })
-        .collect()
-}
-
-/// Common encoded forms of `<` / `>` we look for when deciding whether a
-/// payload depends on angle brackets. A payload that carries any of these
-/// forms is hoping the server single-pass-decodes the input — when the
-/// server filters `<` after decode (the common case), the bypass fails.
-/// Hoisting payloads that carry no angle bracket in any form (event-
-/// handler quote-breakouts, protocol-URI payloads) ahead of these
-/// angle-dependent variants lets the scanner hit a working payload first
-/// and short-circuit the rest of the loop via `reflection_found_locally`.
-const ANGLE_ENCODED_NEEDLES_LT: &[&str] = &[
-    "%3C", "%3c", "%253C", "%253c", "&lt;", "&LT;", "&#60;", "&#x3c;", "&#x3C;", "&#x003c;",
-    "&#x003C;",
-];
-const ANGLE_ENCODED_NEEDLES_GT: &[&str] = &[
-    "%3E", "%3e", "%253E", "%253e", "&gt;", "&GT;", "&#62;", "&#x3e;", "&#x3E;", "&#x003e;",
-    "&#x003E;",
-];
-
-/// True when the payload carries no `<` or `>` in any of: raw bytes,
-/// percent-encoded form (single or double), or HTML entity (named,
-/// decimal, hex). Used to hoist angle-free payloads to the front of the
-/// payload list when `Param.invalid_specials` flags angle brackets — those
-/// payloads (event-handler quote-breakouts, `javascript:` protocol URIs)
-/// are the ones that actually reflect through an angle-stripping filter,
-/// and the loop's `reflection_found_locally` short-circuit means the
-/// first hit zeros out the rest of the budget.
-fn payload_is_angle_free(p: &str) -> bool {
-    if p.contains('<') || p.contains('>') {
-        return false;
-    }
-    for n in ANGLE_ENCODED_NEEDLES_LT {
-        if p.contains(n) {
-            return false;
-        }
-    }
-    for n in ANGLE_ENCODED_NEEDLES_GT {
-        if p.contains(n) {
-            return false;
-        }
-    }
-    true
-}
-
-/// Stable-partition the payload list so payloads that don't depend on
-/// `<`/`>` (in any encoded form) come first when active probing has
-/// flagged angles as invalid. Pairs with [`prune_blocked_raw_angles`]:
-/// pruning kills raw-angle payloads outright, hoisting reorders the
-/// remaining list so the angle-free survivors get tested before the
-/// encoded-angle variants whose only hope is a naive single-pass-decode
-/// filter (rare in practice). Net effect: the first reflection-finding
-/// request usually comes from an angle-free payload, the loop short-
-/// circuits, and the budget for the param collapses from thousands of
-/// requests to dozens.
-fn hoist_angle_free_payloads(payloads: Vec<String>, invalid_specials: &[char]) -> Vec<String> {
-    let block_lt = invalid_specials.contains(&'<');
-    let block_gt = invalid_specials.contains(&'>');
-    if !block_lt && !block_gt {
-        return payloads;
-    }
-    // Three tiers, front to back:
-    //   pad   — leading-window ("positional") bypass payloads (raw `<` on
-    //           purpose). When angles are reported blocked the block may be
-    //           positional, and these are then the *only* shapes that can reach
-    //           a tag injection, so they must lead — otherwise the DOM phase's
-    //           inert-echo early exit can retire before they are ever tried.
-    //   clean — angle-free survivors (event-handler / quote-breakout shapes).
-    //   rest  — encoded-angle variants (need a naive single-pass-decode filter).
-    let mut pad: Vec<String> = Vec::new();
-    let mut clean: Vec<String> = Vec::with_capacity(payloads.len());
-    let mut rest: Vec<String> = Vec::with_capacity(payloads.len());
-    for p in payloads {
-        if crate::payload::synthesis::is_positional_pad_bypass(&p) {
-            pad.push(p);
-        } else if payload_is_angle_free(&p) {
-            clean.push(p);
-        } else {
-            rest.push(p);
-        }
-    }
-    pad.reserve(clean.len() + rest.len());
-    pad.extend(clean);
-    pad.extend(rest);
-    pad
-}
-
-fn get_fallback_reflection_payloads(
-    args: &ScanArgs,
-) -> Result<Vec<String>, Box<dyn std::error::Error>> {
-    let mut base_payloads = vec![];
-
-    if args.only_custom_payload {
-        if let Some(path) = &args.custom_payload {
-            base_payloads.extend(crate::scanning::xss_common::load_custom_payloads(path)?);
-        }
-    } else {
-        // HTML/attribute payloads first — they break out of attribute contexts
-        // and create real DOM elements. JS-only payloads (alert(1), etc.) are
-        // excluded from the reflection list because they cause false-positive R
-        // findings when reflected inside quoted attribute values, blocking the
-        // attribute-breakout payloads that follow.
-        base_payloads.extend(crate::payload::get_dynamic_xss_html_payloads());
-        base_payloads.extend(crate::payload::get_dynamic_xss_attribute_payloads());
-        base_payloads.extend(crate::payload::get_mxss_payloads());
-        base_payloads.extend(crate::payload::get_protocol_injection_payloads());
-        if let Some(path) = &args.custom_payload {
-            base_payloads.extend(crate::scanning::xss_common::load_custom_payloads(path)?);
-        }
-    }
-
-    // Apply encoder policy to unique base payloads
-    let payloads = crate::encoding::apply_encoders_to_payloads(&base_payloads, &args.encoders);
-
-    Ok(payloads)
-}
-
-fn get_js_breakout_payloads() -> Vec<String> {
-    let class_marker = crate::scanning::markers::class_marker();
-    let id_marker = crate::scanning::markers::id_marker();
-
-    let base_templates = [
-        format!("</script><img src=x onerror={{JS}} class={}>", class_marker),
-        format!("</script><svg onload={{JS}} class={}>", class_marker),
-        format!("</script><img src=x onerror={{JS}} id={}>", id_marker),
-    ];
-
-    let breakout_prefixes: &[&str] = &["", "';", "\";", "*/"];
-
-    let mut payloads = Vec::new();
-    for js in crate::payload::XSS_JAVASCRIPT_PAYLOADS_SMALL.iter() {
-        for tmpl in &base_templates {
-            for &prefix in breakout_prefixes {
-                let payload = format!("{}{}", prefix, tmpl.replace("{JS}", js));
-                payloads.push(payload);
-            }
-        }
-    }
-    payloads
-}
-
-/// JSONP-callback payloads: reflected as the *callable identifier* of a
-/// `application/javascript` (JSONP) response — `callback=…` echoed into
-/// `…({"data":1})`. Each is executable JavaScript that calls a visible sink and
-/// comments-out / neutralises the trailing `({…})`, so `has_js_context_evidence`
-/// confirms a real `CallExpression` covered by the payload (the genuine JSONP
-/// XSS the `content_type_is_inert_data` carve-out keeps `application/javascript`
-/// scannable for). On an HTML body these parse-fail as whole-document
-/// JavaScript, so they add no false positives — only the DOM phase's
-/// content-type-aware JS-context check upgrades them to V.
-pub(crate) fn get_jsonp_callback_payloads() -> Vec<String> {
-    vec![
-        "alert(document.domain)//".to_string(),
-        ";alert(1)//".to_string(),
-        "-alert(1)-".to_string(),
-    ]
-}
-
-/// Round-robin interleave several payload families into one list so that any
-/// prefix samples *every* family proportionally (issue #1156). Used for the
-/// unknown-context DOM catalog so the recall-preserving DOM-phase early exit
-/// sees a representative of each DOM-evidence kind before the inert-echo budget
-/// can trip, instead of exhausting the (redundant, sink-varied) HTML-template
-/// block while the protocol/mXSS families sit thousands of payloads away.
-/// Within-family order is preserved; the result contains exactly the union of
-/// the inputs (no dedup — the encoder pass dedups downstream).
-fn interleave_payload_families(families: Vec<Vec<String>>) -> Vec<String> {
-    let total: usize = families.iter().map(Vec::len).sum();
-    // Consume each family (move the Strings out — the generators already handed
-    // us owned clones) rather than re-cloning per element.
-    let mut iters: Vec<_> = families.into_iter().map(Vec::into_iter).collect();
-    let mut out = Vec::with_capacity(total);
-    let mut any = true;
-    while any {
-        any = false;
-        for it in &mut iters {
-            if let Some(p) = it.next() {
-                out.push(p);
-                any = true;
-            }
-        }
-    }
-    out
-}
-
-/// Resolve the effective per-parameter payload cap from the user's
-/// `--max-payloads-per-param` (`max_payloads`) and `--deep-scan` flags.
-///
-/// - An explicit `--max-payloads-per-param N` (N > 0) always wins — it is
-///   honored verbatim even under `--deep-scan`.
-/// - `0` (the default) means "apply the built-in [`DEFAULT_PAYLOAD_SAFETY_CAP`]"
-///   so a parameter that reflects every payload without DOM-verifying can't
-///   drive the DOM phase through the full payload set (issue #1153).
-/// - `--deep-scan` lifts only the built-in safety cap: it makes the default
-///   (`0`) unlimited for exhaustive runs. It does not override an explicit cap.
-pub(crate) fn effective_payload_cap(max_payloads: usize, deep_scan: bool) -> usize {
-    match max_payloads {
-        0 if deep_scan => 0,
-        0 => crate::cmd::scan::DEFAULT_PAYLOAD_SAFETY_CAP,
-        n => n,
-    }
-}
-
-/// Requests one parameter will cost the scan, mirroring `run_scanning`'s own
-/// fan-out: it builds a reflection payload set and a DOM payload set, truncates
-/// **each** to the effective per-parameter cap, and then sends one request per
-/// payload in both (`total_tasks += reflection.len() + dom.len()`).
-///
-/// The DOM half used to be missing here, so `/preflight` quoted roughly half the
-/// requests the scan would send — on the one number the endpoint exists to
-/// produce. `cmd::scan::analysis` already estimated it this way for `--dry-run`;
-/// this brings the REST and MCP endpoints onto the same arithmetic.
-///
-/// Still a lower bound: WAF mutation/encoder expansion and the shared CSP/tech
-/// payloads appended after the cap are not counted, matching the CLI's caveat.
-pub(crate) fn estimate_param_requests(
-    p: &Param,
-    scan_args: &ScanArgs,
-    enc_factor: usize,
-    apply_cap: &dyn Fn(usize) -> usize,
-) -> usize {
-    let refl_len = if let Some(ctx) = &p.injection_context {
-        crate::scanning::xss_common::get_dynamic_payloads(ctx, scan_args)
-            .unwrap_or_else(|_| vec![])
-            .len()
-    } else {
-        let html_len = crate::payload::get_dynamic_xss_html_payloads().len() * enc_factor;
-        let js_len = crate::payload::XSS_JAVASCRIPT_PAYLOADS.len() * enc_factor;
-        html_len + js_len
-    };
-    let dom_len = match &p.injection_context {
-        // A JS-context param gets no DOM-verification pass.
-        Some(crate::parameter_analysis::InjectionContext::Javascript(_)) => 0,
-        Some(ctx) => crate::scanning::xss_common::generate_dynamic_payloads(ctx).len() * enc_factor,
-        None => {
-            (crate::payload::get_dynamic_xss_html_payloads().len()
-                + crate::payload::get_dynamic_xss_attribute_payloads().len())
-                * enc_factor
-        }
-    };
-    apply_cap(refl_len).saturating_add(apply_cap(dom_len))
-}
-
-pub(crate) fn get_dom_payloads(
-    param: &Param,
-    args: &ScanArgs,
-) -> Result<Vec<String>, Box<dyn std::error::Error>> {
-    let mut payloads = get_dom_payloads_for_context(param, args)?;
-
-    // Prepend JSONP-callback verifiers (raw, un-encoded) so a
-    // `application/javascript` endpoint that reflects the callable identifier is
-    // DOM-verified via JS-context evidence — the genuine JSONP XSS — instead of
-    // the false-positive HTML-marker echo the content-type gate now suppresses.
-    // Placed first so they land inside the DOM phase's inert-echo early-exit
-    // window; on an HTML body they parse-fail as whole-document JS and add no
-    // findings. Skipped only when the user restricts to custom payloads.
-    if !args.only_custom_payload {
-        let jsonp = get_jsonp_callback_payloads();
-        if !jsonp.is_empty() {
-            payloads.splice(0..0, jsonp);
-        }
-    }
-    Ok(payloads)
-}
-
-fn get_dom_payloads_for_context(
-    param: &Param,
-    args: &ScanArgs,
-) -> Result<Vec<String>, Box<dyn std::error::Error>> {
-    match &param.injection_context {
-        // JS context: script breakout payloads with markers for DOM verification
-        Some(crate::parameter_analysis::InjectionContext::Javascript(_)) => {
-            let base_payloads = get_js_breakout_payloads();
-            let out = crate::encoding::apply_encoders_to_payloads(&base_payloads, &args.encoders);
-            Ok(out)
-        }
-        // Known non-JS contexts: use locally generated payloads only (exclude remote) to avoid large cross-product
-        Some(ctx) => {
-            // If param has analysis data, use adaptive encoding for better bypass
-            if param.invalid_specials.is_some() || param.valid_specials.is_some() {
-                let invalid = param.invalid_specials.as_deref().unwrap_or_default();
-                let valid = param.valid_specials.as_deref().unwrap_or_default();
-                let payloads =
-                    crate::scanning::xss_common::generate_adaptive_payloads(ctx, invalid, valid);
-                return Ok(payloads);
-            }
-            // Use locally generated payloads only (no remote) to avoid large cross-product in DOM verification
-            let base_payloads = crate::scanning::xss_common::generate_dynamic_payloads(ctx);
-            // Expand with shared encoder policy helper
-            let out = crate::encoding::apply_encoders_to_payloads(&base_payloads, &args.encoders);
-            Ok(out)
-        }
-        // Unknown context: use HTML + Attribute payloads (+ custom if provided), never error
-        None => {
-            // Use only local HTML/Attribute payloads (exclude remote) for DOM verification in unknown contexts
-            let mut base_payloads = vec![];
-
-            if args.only_custom_payload {
-                if let Some(path) = &args.custom_payload {
-                    // Avoid erroring when custom payload file is missing
-                    base_payloads.extend(
-                        crate::scanning::xss_common::load_custom_payloads(path)
-                            .unwrap_or_else(|_| vec![]),
-                    );
-                }
-            } else {
-                // Issue #1156: round-robin interleave the evidence families so
-                // every DOM-evidence kind (HTML-tag, event-handler/attribute,
-                // mXSS, DOM-clobbering, protocol/URL) is represented in any
-                // prefix of the list — in particular within the first
-                // `INERT_ECHO_BUDGET` payloads the DOM-phase early exit samples.
-                // Plain concatenation appended the protocol/mXSS generators last
-                // (thousands of payloads in), so the early exit could cut them
-                // entirely on a raw-reflecting URL-attribute echo, where the
-                // protocol payload is the only verifier. Within-family ordering
-                // (e.g. HTML attribute-breakout templates first) is preserved.
-                base_payloads = interleave_payload_families(vec![
-                    crate::payload::get_dynamic_xss_html_payloads(),
-                    crate::payload::get_dynamic_xss_attribute_payloads(),
-                    crate::payload::get_mxss_payloads(),
-                    crate::payload::get_dom_clobbering_payloads(),
-                    crate::payload::get_protocol_injection_payloads(),
-                ]);
-                if let Some(path) = &args.custom_payload {
-                    base_payloads.extend(
-                        crate::scanning::xss_common::load_custom_payloads(path)
-                            .unwrap_or_else(|_| vec![]),
-                    );
-                }
-            }
-
-            // Ensure we always have DOM-capable payloads for non-JS contexts
-            if base_payloads.is_empty() {
-                base_payloads.extend(crate::payload::get_dynamic_xss_html_payloads());
-                base_payloads.extend(crate::payload::get_dynamic_xss_attribute_payloads());
-            }
-
-            // Expand with shared encoder policy helper
-            let out = crate::encoding::apply_encoders_to_payloads(&base_payloads, &args.encoders);
-            Ok(out)
-        }
-    }
-}
-
-/// Run AST-based DOM XSS static analysis on the given response HTML.
-///
-/// Extracts JavaScript blocks, analyses each for DOM XSS flows, performs
-/// lightweight runtime verification, and returns any findings.  De-duplicates
-/// against `ast_seen` (shared across calls for the same parameter).
-async fn run_ast_dom_analysis(
-    client: &reqwest::Client,
-    target: &Target,
-    param: &Param,
-    response_text: &str,
-    ast_seen: &mut HashSet<String>,
-) -> Vec<crate::scanning::result::Result> {
-    let mut results = Vec::new();
-    let (js_blocks, script_element_ids) =
-        crate::scanning::ast_integration::extract_js_and_script_ids(response_text);
-    let posture = crate::scanning::ast_integration::PageSecurityPosture::from_target(target);
-    for js_code in js_blocks {
-        let findings =
-            crate::scanning::ast_integration::analyze_javascript_for_dom_xss_with_html_context(
-                &js_code,
-                target.url.as_str(),
-                &script_element_ids,
-                posture.trusted_types_enforced,
-            );
-        for (vuln, payload, description) in findings {
-            let self_bootstrap_verified =
-                crate::scanning::ast_integration::has_self_bootstrap_verification(
-                    &js_code,
-                    &vuln.source,
-                );
-            let ast_key = format!(
-                "{}|{}|{}|{}|{}",
-                param.name, vuln.line, vuln.column, vuln.source, vuln.sink
-            );
-            if ast_seen.contains(&ast_key) {
-                continue;
-            }
-            ast_seen.insert(ast_key);
-            let source_uses_url_surface = ast_source_uses_browser_url_surface(&vuln.source);
-            let result_url = if source_uses_url_surface {
-                crate::scanning::ast_integration::build_dom_xss_poc_url(
-                    target.url.as_str(),
-                    &vuln.source,
-                    &payload,
-                )
-            } else {
-                let base = crate::scanning::url_inject::effective_query_base(&target.url, param);
-                crate::scanning::url_inject::build_injected_url(&base, param, &payload)
-            };
-            // Graded from the flow's own shape, so the light-check and
-            // self-bootstrap upgrades below can leave `result_type` at
-            // `Verified` while the grade stays `Low`. See
-            // `ast_integration::build_ast_dom_xss_result` for why that
-            // disagreement is intentional during the tier migration.
-            let (confidence, confidence_reason) =
-                crate::scanning::ast_integration::grade_ast_finding(
-                    &vuln.source,
-                    &vuln.sink,
-                    vuln.guarded,
-                    self_bootstrap_verified,
-                    posture,
-                );
-            let mut ast_result = crate::scanning::result::Result::builder(FindingType::AstDetected)
-                .inject_type("DOM-XSS")
-                .method(crate::scanning::url_inject::effective_method(
-                    &target.method,
-                    param,
-                ))
-                .confidence(confidence, confidence_reason)
-                .data(result_url.clone())
-                .param(param.name.clone())
-                .payload(payload.clone())
-                .evidence(format!(
-                    "{}:{}:{} - {} (Source: {}, Sink: {})",
-                    target.url.as_str(),
-                    vuln.line,
-                    vuln.column,
-                    description,
-                    vuln.source,
-                    vuln.sink
-                ))
-                .cwe("CWE-79")
-                .severity("Medium")
-                .message_id(0)
-                // The confirmation status is appended below, once the light
-                // check has run — a finding this path promotes to `V` must not
-                // also claim it "needs runtime confirmation" (#1238).
-                .message_str(description.clone())
-                .build();
-            ast_result.location = format!("{:?}", param.location);
-            // A URL-surface source got its POC URL from
-            // `build_dom_xss_poc_url`, which already placed the payload in the
-            // fragment / query / path — synthesizing `?param=payload` on top
-            // of that would emit a double-injection POC.
-            ast_result.poc_url_complete = source_uses_url_surface;
-            if !source_uses_url_surface {
-                ast_result.request = Some(build_request_text(target, param, &payload));
-            }
-            ast_result.response = Some(crate::scanning::result::bound_evidence_body(
-                response_text.to_string(),
-                &payload,
-            ));
-            // Lightweight runtime verification (non-headless)
-            let (verified, rt_resp, note) =
-                crate::scanning::light_verify::verify_dom_xss_light_with_client(
-                    client, target, param, &payload,
-                )
-                .await;
-            if let Some(runtime_response) = rt_resp {
-                ast_result.response = Some(crate::scanning::result::bound_evidence_body(
-                    runtime_response,
-                    &payload,
-                ));
-            }
-            if verified {
-                ast_result.result_type = FindingType::Verified;
-                ast_result.severity = "High".to_string();
-                ast_result.message_str =
-                    format!("{} [light check: confirmed]", ast_result.message_str);
-            } else if self_bootstrap_verified {
-                ast_result.result_type = FindingType::Verified;
-                ast_result.severity = "High".to_string();
-                ast_result.message_str = format!(
-                    "{} [static self-bootstrap confirmed]",
-                    ast_result.message_str
-                );
-            } else {
-                ast_result.message_str = format!(
-                    "{} (needs runtime confirmation) [light check: not confirmed]",
-                    ast_result.message_str
-                );
-            }
-            // Appended last so the CSP-style caveat trails the verdict instead
-            // of splitting the description from it.
-            if let Some(n) = note {
-                ast_result.message_str = format!("{} [{}]", ast_result.message_str, n);
-            }
-            results.push(ast_result);
-        }
-    }
-    results
-}
-
-/// Append a batch of findings to the shared results vector and bump the
-/// running findings counter. No-op when `batch` is empty. Centralizes the
-/// lock + extend + counter-update sequence shared by every preflight finding
-/// source (libs, initial AST, external JS) across the CLI, server, and MCP
-/// surfaces.
-///
-/// The counter is bumped by the number of findings that match
-/// `limit_result_type` (already-uppercased `--limit-result-type`), mirroring
-/// [`ScanWorkerCtx::flush_results`] — otherwise N preflight findings of a
-/// non-matching type would trip `--limit N` and short-circuit the injection
-/// phase before any matching finding is produced.
-pub(crate) async fn accumulate_findings(
-    results: &tokio::sync::Mutex<Vec<crate::scanning::result::Result>>,
-    findings_count: &std::sync::atomic::AtomicUsize,
-    batch: Vec<crate::scanning::result::Result>,
-    limit_result_type: &str,
-) {
-    if batch.is_empty() {
-        return;
-    }
-    let added = count_matching_results(&batch, limit_result_type);
-    results.lock().await.extend(batch);
-    findings_count.fetch_add(added, std::sync::atomic::Ordering::Relaxed);
-}
-
-/// Fetch all same-origin `<script src>` bundles referenced by `html` and run
-/// AST DOM-XSS analysis on each one. Called once per target at the pre-scan
-/// (preflight) stage so it fires even for SPAs that have no server-side
-/// parameter reflection (where the per-param probe loop never executes).
-///
-/// Returns an empty `Vec` when `--analyze-external-js` is not set.
-pub(crate) async fn fetch_and_analyze_external_js(
-    client: &reqwest::Client,
-    target: &Target,
-    html: &str,
-    scan_args: &ScanArgs,
-) -> Vec<crate::scanning::result::Result> {
-    if !scan_args.analyze_external_js {
-        return Vec::new();
-    }
-
-    // Compile scope filters once rather than per-URL.
-    let include_patterns: Vec<regex::Regex> = scan_args
-        .include_url
-        .iter()
-        .filter_map(|p| regex::Regex::new(p).ok())
-        .collect();
-    let exclude_patterns: Vec<regex::Regex> = scan_args
-        .exclude_url
-        .iter()
-        .filter_map(|p| regex::Regex::new(p).ok())
-        .collect();
-
-    let script_urls =
-        crate::scanning::ast_integration::extract_same_origin_script_srcs(html, &target.url);
-
-    let script_element_ids = crate::scanning::ast_integration::extract_script_element_ids(html);
-    // The posture comes from the *page's* CSP, not each script's response: the
-    // policy that governs whether an injected handler runs is the document's.
-    let posture = crate::scanning::ast_integration::PageSecurityPosture::from_target(target);
-    let trusted_types_enforced = posture.trusted_types_enforced;
-    let mut results: Vec<crate::scanning::result::Result> = Vec::new();
-
-    // extract_same_origin_script_srcs already deduplicates; just cap the count.
-    for script_url in script_urls.into_iter().take(MAX_EXTERNAL_JS_FILES) {
-        let url_str = script_url.as_str().to_owned();
-
-        // Apply --include-url / --exclude-url scope to external script URLs.
-        if !include_patterns.is_empty() && !include_patterns.iter().any(|r| r.is_match(&url_str)) {
-            continue;
-        }
-        if exclude_patterns.iter().any(|r| r.is_match(&url_str)) {
-            continue;
-        }
-
-        let rb =
-            crate::utils::build_request(client, target, reqwest::Method::GET, script_url, None);
-        let send_result =
-            crate::utils::send_with_retry(rb, scan_args.retries, scan_args.retry_delay).await;
-        // Count this external-JS fetch (up to MAX_EXTERNAL_JS_FILES per page);
-        // its retries, if any, are counted inside send_with_retry. These GETs
-        // were previously missing from REQUEST_COUNT / the live req/s rate.
-        crate::tick_request_count();
-        let resp = match send_result {
-            Ok(r) => r,
-            Err(_) => continue,
-        };
-        if !resp.status().is_success() {
-            continue;
-        }
-        let body = match crate::utils::http::read_body(resp).await {
-            Ok(b) => b,
-            Err(_) => continue,
-        };
-        if body.len() > MAX_EXTERNAL_JS_BYTES {
-            continue;
-        }
-
-        let findings =
-            crate::scanning::ast_integration::analyze_javascript_for_dom_xss_with_html_context(
-                &body,
-                target.url.as_str(),
-                &script_element_ids,
-                trusted_types_enforced,
-            );
-
-        for (vuln, payload, description) in findings {
-            let self_bootstrap_verified =
-                crate::scanning::ast_integration::has_self_bootstrap_verification(
-                    &body,
-                    &vuln.source,
-                );
-            let message =
-                format!("{description} (needs runtime confirmation) [external JS: {url_str}]");
-            let evidence = format!(
-                "{}:{}:{} - {} (Source: {}, Sink: {}) [script: {}]",
-                target.url.as_str(),
-                vuln.line,
-                vuln.column,
-                description,
-                vuln.source,
-                vuln.sink,
-                url_str,
-            );
-            results.push(crate::scanning::ast_integration::build_ast_dom_xss_result(
-                crate::scanning::ast_integration::AstDomFinding {
-                    target_url: target.url.as_str(),
-                    target_method: &target.method,
-                    vuln: &vuln,
-                    payload,
-                    evidence,
-                    message,
-                    self_bootstrap_verified,
-                    posture,
-                },
-            ));
-        }
-    }
-
-    results
-}
-
-fn build_request_text(target: &Target, param: &Param, payload: &str) -> String {
-    use crate::parameter_analysis::Location;
-    let url = match param.location {
-        Location::Query => {
-            // Show the request against the actual sink URL — form action when
-            // the param came from form discovery, otherwise target.url. The
-            // displayed PoC must match the URL that scanning actually hits.
-            let base = crate::scanning::url_inject::effective_query_base(&target.url, param);
-            let mut pairs: Vec<(String, String)> = base
-                .query_pairs()
-                .map(|(k, v)| (k.to_string(), v.to_string()))
-                .collect();
-            let mut found = false;
-            for pair in &mut pairs {
-                if pair.0 == param.name {
-                    pair.1 = payload.to_string();
-                    found = true;
-                    break;
-                }
-            }
-            if !found {
-                pairs.push((param.name.clone(), payload.to_string()));
-            }
-            let query = url::form_urlencoded::Serializer::new(String::new())
-                .extend_pairs(&pairs)
-                .finish();
-            let mut url = base;
-            url.set_query(Some(&query));
-            url
-        }
-        Location::Path => {
-            // Inject into a specific path segment (param.name pattern: path_segment_{idx})
-            let mut url = target.url.clone();
-            if let Some(idx_str) = param.name.strip_prefix("path_segment_")
-                && let Ok(idx) = idx_str.parse::<usize>()
-            {
-                let original_path = url.path();
-                let mut segments: Vec<&str> = if original_path == "/" {
-                    Vec::new()
-                } else {
-                    original_path
-                        .trim_matches('/')
-                        .split('/')
-                        .filter(|s| !s.is_empty())
-                        .collect()
-                };
-                if idx < segments.len() {
-                    segments[idx] = payload;
-                    let new_path = if segments.is_empty() {
-                        "/".to_string()
-                    } else {
-                        format!("/{}", segments.join("/"))
-                    };
-                    url.set_path(&new_path);
-                }
-            }
-            url
-        }
-        Location::Body
-        | Location::JsonBody
-        | Location::MultipartBody
-        | Location::GraphqlBody
-        | Location::XmlBody => {
-            // Body params use the form action URL when discovered from a form,
-            // so the displayed request matches the POST actually sent.
-            crate::scanning::url_inject::effective_query_base(&target.url, param)
-        }
-        _ => target.url.clone(),
-    };
-
-    let method = crate::scanning::url_inject::effective_method(&target.method, param);
-    // Body-bearing locations always send a body; synthesize one when the
-    // target has no original `data`, so the displayed PoC isn't an empty POST.
-    let (body, content_type): (Option<String>, Option<String>) = match param.location {
-        Location::Body => {
-            let body = crate::scanning::url_inject::urlencoded_body(
-                target.data.as_deref(),
-                &param.name,
-                payload,
-            );
-            (
-                Some(body),
-                Some("application/x-www-form-urlencoded".to_string()),
-            )
-        }
-        Location::JsonBody => {
-            let body = crate::scanning::url_inject::json_body(
-                target.data.as_deref(),
-                &param.name,
-                &param.value,
-                payload,
-            );
-            (Some(body), Some("application/json".to_string()))
-        }
-        Location::MultipartBody => (target.data.clone(), Some("multipart/form-data".to_string())),
-        // GraphQL / XML rebuild the whole body from the param's pipeline
-        // (`JsonField` into the GraphQL request / `Splice` around the XML
-        // injection point). `apply_param_encoding` runs that pipeline on the
-        // raw `payload`, so the displayed PoC body is exactly what goes on the
-        // wire.
-        Location::GraphqlBody => {
-            let body = crate::encoding::pre_encoding::apply_param_encoding(payload, param);
-            (Some(body), Some("application/json".to_string()))
-        }
-        Location::XmlBody => {
-            let body = crate::encoding::pre_encoding::apply_param_encoding(payload, param);
-            (
-                Some(body),
-                Some(crate::scanning::url_inject::xml_request_content_type(
-                    target,
-                )),
-            )
-        }
-        _ => (target.data.clone(), None),
-    };
-
-    let mut buf = String::with_capacity(512);
-
-    // Request line
-    buf.push_str(&method);
-    buf.push(' ');
-    buf.push_str(url.path());
-    if let Some(q) = url.query() {
-        buf.push('?');
-        buf.push_str(q);
-    }
-    buf.push_str(" HTTP/1.1\r\nHost: ");
-    buf.push_str(url.host_str().unwrap_or(""));
-    // `Host` is host *and port*. Dropping it made every PoC against a
-    // non-default port replay to :80/:443 — a different service, or nothing.
-    // `Url::port()` is already `None` for the scheme's default port, which is
-    // exactly when the Host header should omit it.
-    if let Some(port) = url.port() {
-        buf.push(':');
-        buf.push_str(&port.to_string());
-    }
-
-    // A `Location::Header` param is injected into the request itself, not the
-    // URL — as a header of its own name, or, when the name is one of the
-    // target's cookies, into `Cookie` (see `url_inject::build_header_request`,
-    // which this must mirror or the PoC does not reproduce what was sent).
-    let header_param = matches!(param.location, Location::Header);
-    let cookie_param = header_param && crate::scanning::url_inject::param_is_cookie(target, param);
-    let injected_header = header_param && !cookie_param;
-
-    for (k, v) in &target.headers {
-        // `apply_header_overrides` replaces rather than appends, so a
-        // same-named original is not on the wire and must not be shown here.
-        if injected_header && k.eq_ignore_ascii_case(&param.name) {
-            continue;
-        }
-        // Same for the composed `Cookie` a cookie param is injected through
-        // (`build_request_with_cookie` overrides any captured Cookie header),
-        // and for a captured `Content-Type` on a body injection, which
-        // `build_body_request_base` drops in favour of the injector's.
-        if cookie_param && k.eq_ignore_ascii_case("cookie") {
-            continue;
-        }
-        if content_type.is_some() && k.eq_ignore_ascii_case("content-type") {
-            continue;
-        }
-        buf.push_str("\r\n");
-        buf.push_str(k);
-        buf.push_str(": ");
-        buf.push_str(v);
-    }
-    if injected_header {
-        buf.push_str("\r\n");
-        buf.push_str(&param.name);
-        buf.push_str(": ");
-        buf.push_str(payload);
-    }
-    // Body injectors always set their own `Content-Type` on the wire; a
-    // Query/Path injector re-sends the original body verbatim and passes
-    // `None` here, keeping whatever the target captured.
-    if let Some(ct) = &content_type {
-        buf.push_str("\r\nContent-Type: ");
-        buf.push_str(ct);
-    }
-
-    // Cookies. For a cookie param the injected value is written first and the
-    // target's other cookies follow, matching `build_header_request` — those
-    // neighbours often carry the session state the sink needs to render at all.
-    if cookie_param {
-        buf.push_str("\r\nCookie: ");
-        buf.push_str(&param.name);
-        buf.push('=');
-        buf.push_str(payload);
-        if let Some(rest) =
-            crate::utils::compose_cookie_header_excluding(&target.cookies, Some(&param.name))
-            && !rest.is_empty()
-        {
-            buf.push_str("; ");
-            buf.push_str(&rest);
-        }
-    } else if !target.cookies.is_empty()
-        && !target
-            .headers
-            .iter()
-            .any(|(k, _)| k.eq_ignore_ascii_case("cookie"))
-    {
-        // Mirrors `apply_headers_ua_cookies`: `target.cookies` are auto-attached
-        // only when the target does not already carry its own Cookie header
-        // (already emitted above), never as a second Cookie line.
-        buf.push_str("\r\nCookie: ");
-        for (i, (k, v)) in target.cookies.iter().enumerate() {
-            if i > 0 {
-                buf.push_str("; ");
-            }
-            buf.push_str(k);
-            buf.push('=');
-            buf.push_str(v);
-        }
-    }
-
-    if let Some(data) = &body {
-        buf.push_str("\r\nContent-Length: ");
-        buf.push_str(&data.len().to_string());
-        buf.push_str("\r\n\r\n");
-        buf.push_str(data);
-    } else {
-        buf.push_str("\r\n");
-    }
-
-    buf
-}
-
-fn ast_source_uses_browser_url_surface(source: &str) -> bool {
-    source.contains("location.hash")
-        || source.contains("location.search")
-        || source.contains("URLSearchParams.get(")
-        || source.contains("location.href")
-        || source.contains("location.pathname")
-        || source.contains("document.URL")
-        || source.contains("window.opener")
-        || source.contains("event.newValue")
-        || source.contains("event.oldValue")
-}
-
-/// Compute the WAF bypass strategy for this scan. Returns `None` when WAF
-/// bypass is disabled (`--waf-bypass off`), no WAF was fingerprinted, or the
-/// fingerprint set is empty — in which case payload generation skips the
-/// mutation / extra-encoder expansion entirely.
-fn compute_waf_strategy(
-    target: &Target,
-    args: &ScanArgs,
-) -> Option<crate::waf::bypass::BypassStrategy> {
-    if args.waf_bypass == "off" {
-        return None;
-    }
-    target.waf_info.as_ref().and_then(|waf_info| {
-        if waf_info.is_empty() {
-            None
-        } else {
-            let waf_types: Vec<&crate::waf::WafType> = waf_info.waf_types();
-            Some(crate::waf::bypass::merge_strategies(&waf_types))
-        }
-    })
-}
-
-/// Find a CSP declared with `<meta http-equiv>`, returning the equivalent
-/// header name and the policy text. Pages served without a CSP header commonly
-/// carry one this way, and the CLI preflight has always honoured it — this is
-/// that logic, shared so the server / MCP surfaces cannot analyse a different
-/// policy than the CLI does for the same page.
-pub(crate) fn extract_meta_csp(html: &str) -> Option<(String, String)> {
-    let doc = crate::utils::html::parse_document_bounded(html);
-    // Prefer an ENFORCING `Content-Security-Policy` meta over a report-only one,
-    // mirroring the header path's enforcing-over-report-only `.or_else`. A page
-    // may carry both (report-only for telemetry, enforcing for protection), and
-    // their document order is arbitrary — returning whichever appears first
-    // could hand back the report-only policy, which downstream marks
-    // `report_only = true` and zeroes `require_trusted_types_for`. That drops
-    // the real enforcing policy: a TT-hardened enforcing meta would then be
-    // ignored and its (neutralised) DOM findings surface as false positives.
-    let mut report_only: Option<(String, String)> = None;
-    for el in doc.select(crate::scanning::selectors::meta_csp()) {
-        let http_equiv = el
-            .value()
-            .attr("http-equiv")
-            .unwrap_or("")
-            .to_ascii_lowercase();
-        let content = el.value().attr("content").unwrap_or("");
-        if content.is_empty() {
-            continue;
-        }
-        match http_equiv.as_str() {
-            "content-security-policy" => {
-                // Enforcing policy wins outright.
-                return Some(("Content-Security-Policy".to_string(), content.to_string()));
-            }
-            "content-security-policy-report-only" => {
-                // Remember the first report-only, but keep scanning for an
-                // enforcing policy which takes precedence.
-                report_only.get_or_insert_with(|| {
-                    (
-                        "Content-Security-Policy-Report-Only".to_string(),
-                        content.to_string(),
-                    )
-                });
-            }
-            _ => continue,
-        }
-    }
-    report_only
-}
-
-/// Whether the response headers carry an **enforcing** CSP with
-/// `require-trusted-types-for 'script'`. Used by the server / MCP surfaces —
-/// which fetch the page directly rather than through the preflight stage — to
-/// give the initial AST DOM analysis the same Trusted Types awareness the CLI
-/// path derives from `target.csp_analysis`.
-///
-/// The report-only variant (`Content-Security-Policy-Report-Only`) is
-/// deliberately ignored: it only emits violation reports and enforces nothing,
-/// so the browser does not route sinks through the default policy. Treating it
-/// as enforcement would suppress genuine findings (a false negative).
-pub fn csp_requires_trusted_types(headers: &reqwest::header::HeaderMap) -> bool {
-    headers
-        .get("content-security-policy")
-        .and_then(|v| v.to_str().ok())
-        .map(|v| crate::payload::xss_csp_bypass::analyze_csp(v).require_trusted_types_for)
-        .unwrap_or(false)
-}
-
-/// Pre-merge the payloads shared across every parameter: CSP-bypass payloads
-/// (when CSP was analysed) followed by technology-specific payloads (when a
-/// stack was fingerprinted). Built once per scan so [`generate_param_jobs`]
-/// can clone the merged set per parameter instead of recomputing it.
-fn build_shared_payloads(target: &Target) -> Vec<String> {
-    let csp_bypass_payloads: Vec<String> = target
-        .csp_analysis
-        .as_ref()
-        .map(crate::payload::xss_csp_bypass::get_csp_bypass_payloads)
-        .unwrap_or_default();
-    let tech_payloads: Vec<String> = target
-        .tech_info
-        .as_ref()
-        .map(crate::scanning::tech_detect::get_tech_specific_payloads)
-        .unwrap_or_default();
-    let mut shared = Vec::with_capacity(csp_bypass_payloads.len() + tech_payloads.len());
-    shared.extend(csp_bypass_payloads);
-    shared.extend(tech_payloads);
-    shared
-}
-
-/// Expand a parameter's base payload set into its WAF-bypass variants,
-/// keeping the two bypass axes orthogonal instead of multiplying them.
-///
-/// Output order (front to back), de-duplicated while preserving first
-/// occurrence:
-///   1. the originals — cheapest, browser-native; they reflect first so a
-///      param whose base shape isn't filtered short-circuits immediately;
-///   2. raw structural mutations (`<scr<!---->ipt>`, `<ScRiPt>`, …) — the
-///      highest-probability WAF bypass that needs no server-side decode,
-///      front-loaded so an actively-blocking WAF surfaces a working bypass
-///      before we spend requests on the heavier encoder variants;
-///   3. encoder variants of the originals (`%3C…`, fullwidth, zwsp, …) —
-///      transport-style evasion that relies on the app decoding the wire
-///      bytes back into an executable payload.
-///
-/// Crucially this does *not* emit `encode(mutate(p))`: cross-encoding a
-/// structural mutation buries its bypass under transport encoding and needs
-/// both an app-side decode *and* browser tolerance of the mutated shape — a
-/// compound condition that rarely lands while costing one request apiece.
-/// Skipping it shrinks the per-param request count from `N·(1+m)·(1+k)` to
-/// `N·(1+m+k)` with no loss of reach on either axis.
-///
-/// `stats`, when present, records each generated mutation variant against
-/// its `MutationType` for `target_summary.waf.bypass.mutations_applied[]`
-/// (counted pre-dedup against the encoder set, matching the prior
-/// "did the mutation apply" semantics).
-fn expand_waf_payloads(
-    base: &[String],
-    strategy: &crate::waf::bypass::BypassStrategy,
-    stats: Option<&crate::waf::bypass::MutationStats>,
-) -> Vec<String> {
-    // Size both collections to the orthogonal output estimate
-    // (`N·(1+m+k)`) so the mutation/encoder passes don't repeatedly rehash
-    // and realloc — `base.len()` alone under-allocates by that whole factor.
-    let est = base
-        .len()
-        .saturating_mul(1 + strategy.mutations.len() + strategy.extra_encoders.len());
-    let mut seen: HashSet<String> = HashSet::with_capacity(est);
-    let mut out: Vec<String> = Vec::with_capacity(est);
-
-    // 1. Originals (de-duplicated), kept at the front.
-    for p in base {
-        if seen.insert(p.clone()) {
-            out.push(p.clone());
-        }
-    }
-
-    // 2. Raw structural mutations — no transport encoding applied.
-    if !strategy.mutations.is_empty() {
-        let tagged = crate::waf::bypass::apply_mutations_tagged(
-            base,
-            &strategy.mutations,
-            MAX_WAF_MUTATION_VARIANTS_PER_PAYLOAD,
-        );
-        for (p, origin) in tagged {
-            // `None` is the unmodified base, already emitted in step 1.
-            if let Some(m) = origin {
-                if let Some(stats) = stats {
-                    stats.record_variant(m);
-                }
-                if seen.insert(p.clone()) {
-                    out.push(p);
-                }
-            }
-        }
-    }
-
-    // 3. Encoder variants of the originals. `apply_encoders_to_payloads`
-    //    re-emits each original as the first variant of its base; those
-    //    collide with step 1 and are dropped by `seen`, leaving only the
-    //    genuinely encoded forms.
-    if !strategy.extra_encoders.is_empty() {
-        for v in crate::encoding::apply_encoders_to_payloads(base, &strategy.extra_encoders) {
-            if seen.insert(v.clone()) {
-                out.push(v);
-            }
-        }
-    }
-
-    out
-}
-
-/// === Stage 4: Payload Generation — build per-parameter payload sets ===
-///
-/// For each (non-fragment) reflection parameter, build a [`ParamPayloadJob`]
-/// of `(param, reflection payloads, DOM payloads)` by applying, in order:
-/// context-aware base generation, the shared CSP/tech payloads, WAF bypass
-/// mutations + extra encoders, the adaptive angle prune/hoist, and the
-/// `--max-payloads-per-param` cap. Returns the jobs plus the total payload
-/// count used to size the progress bar (one tick per reflection + DOM
-/// payload).
-/// Whether the HTTP scan phase will actually test `param`.
-///
-/// URL fragments are client-side only — HTTP servers never see them — so
-/// `generate_param_jobs` spawns no worker for `Location::Fragment` params (the
-/// AST DOM analyzer detects `location.hash` sources from response JS
-/// independently, so coverage isn't lost). Async front-ends use this so their
-/// `params_total` matches the number of per-parameter workers actually run
-/// (keeping `estimated_completion_pct` honest) and so preflight estimates don't
-/// bill requests for params that send none.
-pub fn param_is_http_scannable(param: &crate::parameter_analysis::Param) -> bool {
-    !matches!(
-        param.location,
-        crate::parameter_analysis::Location::Fragment
-    )
-}
-
-/// Count of a target's parameters the HTTP scan phase will actually test.
-pub fn http_scannable_param_count(target: &Target) -> usize {
-    target
-        .reflection_params
-        .iter()
-        .filter(|p| param_is_http_scannable(p))
-        .count()
-}
-
-fn generate_param_jobs(
-    target: &Target,
-    args: &ScanArgs,
-    waf_strategy: Option<&crate::waf::bypass::BypassStrategy>,
-    shared_payloads: &[String],
-) -> (Vec<ParamPayloadJob>, u64) {
-    let mut total_tasks = 0u64;
-    let mut param_jobs: Vec<ParamPayloadJob> = Vec::with_capacity(target.reflection_params.len());
-    for param in &target.reflection_params {
-        // URL fragments are client-side only — HTTP servers never see them, so
-        // reflection probes for `Location::Fragment` params would be pure-waste
-        // requests. See `param_is_http_scannable`.
-        if !param_is_http_scannable(param) {
-            continue;
-        }
-        let mut reflection_payloads = if let Some(context) = &param.injection_context {
-            crate::scanning::xss_common::get_dynamic_payloads(context, args)
-                .unwrap_or_else(|_| vec![])
-        } else {
-            get_fallback_reflection_payloads(args).unwrap_or_else(|_| vec![])
-        };
-        let mut dom_payloads = get_dom_payloads(param, args).unwrap_or_else(|_| vec![]);
-
-        // Issue #1075: prepend filter-constrained synthesized payloads to the
-        // reflection set when active probing produced a character profile for
-        // this parameter. (Non-JS DOM payloads receive synthesis separately
-        // inside `get_dom_payloads` → `generate_adaptive_payloads`; JS-context
-        // DOM payloads use the dedicated script-breakout set instead.) Placing
-        // them first lets the first-hit-wins reflection loop try shapes built
-        // for this exact filter before the broad catalog, lifting detection on
-        // custom filters; under non-`--deep-scan` runs that ordering can also
-        // cut requests by hitting earlier.
-        //
-        // Note: with an explicit small `--max-payloads-per-param` (< the synth
-        // count), the truncation below can evict the catalog entirely in favour
-        // of these higher-signal synthesized payloads — intentional, since the
-        // user asked for few payloads and these are the ones most likely to fire.
-        if let Some(context) = &param.injection_context
-            && (param.invalid_specials.is_some() || param.valid_specials.is_some())
-        {
-            let invalid = param.invalid_specials.as_deref().unwrap_or_default();
-            let valid = param.valid_specials.as_deref().unwrap_or_default();
-            // #1072: escaped-quote signal (JS string contexts) drives synthesis
-            // to emit backslash-prefixed breakouts that survive server escaping.
-            let escaped = param.escaped_specials.as_deref().unwrap_or_default();
-            // #1073 follow-up: the breakout computed from this site's observed
-            // inline-<script> prefix, emitted ahead of the fixed catalog.
-            let observed_breakout = param.js_breakout.as_deref();
-            let synthesized = crate::payload::synthesis::synthesize_payloads(
-                context,
-                invalid,
-                valid,
-                escaped,
-                observed_breakout,
-            );
-            if !synthesized.is_empty() {
-                let mut seen: std::collections::HashSet<String> =
-                    std::collections::HashSet::with_capacity(
-                        synthesized.len() + reflection_payloads.len(),
-                    );
-                let mut merged = Vec::with_capacity(synthesized.len() + reflection_payloads.len());
-                for p in synthesized.into_iter().chain(reflection_payloads) {
-                    if seen.insert(p.clone()) {
-                        merged.push(p);
-                    }
-                }
-                reflection_payloads = merged;
-            }
-        }
-
-        // Cap the *base* reflection and DOM catalogs independently (each to
-        // `cap`) BEFORE WAF-bypass expansion and the shared-payload append below.
-        // An explicit --max-payloads-per-param wins; otherwise a built-in safety
-        // cap bounds the request fan-out on parameters that reflect every payload
-        // without DOM-verifying (self-link echoes — issue #1153), where the DOM
-        // phase would otherwise send the full ~10k+ payload set. --deep-scan lifts
-        // the built-in cap (truly unlimited).
-        //
-        // Capping the BASE catalog (not the post-expansion set) is deliberate:
-        // `expand_waf_payloads` keeps the de-duplicated originals at the front and
-        // appends every structural-mutation / extra-encoder variant at the tail,
-        // so capping *after* expansion truncates 100% of the WAF-bypass variants
-        // whenever the base alone exceeds the cap (e.g. ~9k attribute-context
-        // payloads vs the 3000 default) — silently defeating WAF bypass on exactly
-        // the params it was selected for. Capping first lets each surviving base
-        // payload keep its full bypass expansion; with no WAF the set is unchanged
-        // by expansion, so the self-link amplification is still bounded to `cap`.
-        let cap = effective_payload_cap(args.max_payloads_per_param, args.deep_scan);
-        if cap > 0 {
-            let refl_dropped = reflection_payloads.len().saturating_sub(cap);
-            let dom_dropped = dom_payloads.len().saturating_sub(cap);
-            reflection_payloads.truncate(cap);
-            dom_payloads.truncate(cap);
-            // Surface the built-in cap (an explicit user cap is the operator's own
-            // deliberate choice, not a silent one) so a trimmed run is never
-            // silent — "No silent caps".
-            if args.max_payloads_per_param == 0 && (refl_dropped > 0 || dom_dropped > 0) {
-                crate::dbg_log!(
-                    "param {}: built-in payload safety cap {} trimmed {} reflection + {} DOM base payload(s); raise with --max-payloads-per-param or --deep-scan",
-                    param.name,
-                    cap,
-                    refl_dropped,
-                    dom_dropped
-                );
-            }
-        }
-
-        // Apply WAF bypass expansion if a WAF was detected. The two bypass
-        // axes are kept orthogonal rather than multiplied together (see
-        // `expand_waf_payloads`): structural mutations are sent raw and
-        // encoder variants are applied to the originals only, so we never
-        // emit the low-yield `encode(mutate(p))` cross product. The tagged
-        // mutation pass still feeds `record_variant`, which powers the
-        // per-target effectiveness counter in
-        // target_summary.waf.bypass.mutations_applied[].
-        if let Some(strategy) = waf_strategy {
-            let stats = target.mutation_stats.as_deref();
-            reflection_payloads = expand_waf_payloads(&reflection_payloads, strategy, stats);
-            dom_payloads = expand_waf_payloads(&dom_payloads, strategy, stats);
-        }
-
-        // Adaptive prune + reorder: when active probing recorded that the
-        // server strips `<` / `>`, (a) drop reflection/DOM payloads whose
-        // raw bytes carry those characters (guaranteed misses), and
-        // (b) hoist payloads that carry no `<`/`>` in any encoded form to
-        // the front of the list. The reorder is the bigger lever: the
-        // loop's `reflection_found_locally` short-circuit means the first
-        // reflecting payload zeros out the rest of the budget, so putting
-        // angle-free payloads (event-handler / quote-breakout shapes that
-        // actually work against an angle-stripping filter) before the
-        // angle-encoded variants collapses the per-param request count
-        // from thousands to dozens on attribute-context params.
-        if let Some(invalid) = param.invalid_specials.as_deref()
-            && !invalid.is_empty()
-        {
-            let refl_before = reflection_payloads.len();
-            let dom_before = dom_payloads.len();
-            reflection_payloads = prune_blocked_raw_angles(reflection_payloads, invalid);
-            dom_payloads = prune_blocked_raw_angles(dom_payloads, invalid);
-            reflection_payloads = hoist_angle_free_payloads(reflection_payloads, invalid);
-            dom_payloads = hoist_angle_free_payloads(dom_payloads, invalid);
-            if refl_before != reflection_payloads.len() || dom_before != dom_payloads.len() {
-                crate::dbg_log!(
-                    "adaptive prune (param={}): reflection {}→{}, dom {}→{} (invalid_specials={:?})",
-                    param.name,
-                    refl_before,
-                    reflection_payloads.len(),
-                    dom_before,
-                    dom_payloads.len(),
-                    invalid,
-                );
-            }
-        }
-
-        // Append shared payloads (CSP bypass + tech-specific) AFTER the cap so
-        // the safety cap can never trim these few, high-value payloads. They
-        // still get the same WAF-bypass expansion as the base set.
-        if !shared_payloads.is_empty() {
-            let mut shared_refl: Vec<String> = shared_payloads.to_vec();
-            let mut shared_dom: Vec<String> = shared_payloads.to_vec();
-            if let Some(strategy) = waf_strategy {
-                let stats = target.mutation_stats.as_deref();
-                shared_refl = expand_waf_payloads(&shared_refl, strategy, stats);
-                shared_dom = expand_waf_payloads(&shared_dom, strategy, stats);
-            }
-            // Apply the same adaptive prune the base set got: when the server
-            // strips `<`/`>`, angle-bearing shared payloads (many CSP-bypass
-            // shapes) are guaranteed misses, so drop them rather than spend a
-            // request each. Preserves the pre-cap-reorder behavior for shared.
-            if let Some(invalid) = param.invalid_specials.as_deref()
-                && !invalid.is_empty()
-            {
-                shared_refl = prune_blocked_raw_angles(shared_refl, invalid);
-                shared_dom = prune_blocked_raw_angles(shared_dom, invalid);
-            }
-            reflection_payloads.extend(shared_refl);
-            dom_payloads.extend(shared_dom);
-        }
-
-        // One pb.inc(1) per reflection payload plus one per DOM payload.
-        // The previous `len * (1 + len)` formula overcounted by orders of
-        // magnitude, which made `{eta}` meaningless (it would project hours
-        // for a sub-minute scan).
-        total_tasks += reflection_payloads.len() as u64 + dom_payloads.len() as u64;
-        param_jobs.push((param.clone(), reflection_payloads, dom_payloads));
-    }
-    (param_jobs, total_tasks)
-}
-
-/// Build the per-target indicatif progress bar (one tick per reflection /
-/// DOM payload). Returns `None` when no `MultiProgress` is supplied (quiet /
-/// embedded runs), in which case the scan loop simply skips the `inc(1)`
-/// calls.
-fn build_scan_progress_bar(
-    multi_pb: &Option<Arc<MultiProgress>>,
-    total_tasks: u64,
-    target: &Target,
-) -> Option<ProgressBar> {
-    let mp = multi_pb.as_ref()?;
-    let pb = mp.add(ProgressBar::new(total_tasks));
-    // `{per_sec}` would measure pb-position rate, not HTTP request rate;
-    // many `pb.inc(1)` calls here are "no-op" iterations (param already
-    // found, payload skipped), which inflated the rate (e.g. 11.5k/s on
-    // a 5k-payload scan that finished in 0.4s). See req_per_sec_tracker
-    // for the displayed semantics and caveats.
-    //
-    // The trailing `{wave}` paints "Scanning <url>" with the shared metallic
-    // shimmer, re-evaluated on every steady tick from the bar's elapsed time
-    // (no extra timer task). `finish_scan_bar` swaps in a `{msg}` style at the
-    // end so the completion line replaces the wave instead of duplicating it.
-    let req_start = crate::REQUEST_COUNT.load(Ordering::Relaxed);
-    pb.set_style(
-        ProgressStyle::default_bar()
-            .template(
-                "{spinner:.cyan} [{elapsed_precise}] [{bar:28.45/238}] {pos:>5}/{len:5} · {req_per_sec} · {wave}",
-            )
-            .expect("valid progress bar template")
-            .tick_chars(crate::utils::shimmer::TICK_CHARS)
-            .with_key("req_per_sec", req_per_sec_tracker(req_start))
-            .with_key(
-                "wave",
-                crate::utils::shimmer::wave_tracker(
-                    format!("Scanning {}", target.url),
-                    crate::utils::shimmer::BAR_WAVE_RESERVE,
-                ),
-            )
-            .progress_chars("█▉▊▋▌▍▎▏░"),
-    );
-    pb.enable_steady_tick(Duration::from_millis(
-        crate::utils::shimmer::FRAME_MS as u64,
-    ));
-    Some(pb)
-}
-
-/// Render `pb` in its terminal "done" state and stop it.
-///
-/// The in-progress style paints the label through a `{wave}` shimmer and has
-/// no `{msg}` slot, so a plain `finish_with_message` would never show the
-/// completion text. Swap in a compact `{prefix} [elapsed] {msg}` style first:
-/// a status glyph (green `✓` / yellow `⚠`), the elapsed time, then the final
-/// message — replacing the wave rather than printing alongside it.
-pub(crate) fn finish_scan_bar(pb: &ProgressBar, prefix: String, msg: String) {
-    // Keep the completion line on one row too. The finished template is
-    // `{prefix} [elapsed] {msg}` — ~13 cols of furniture before `{msg}` — so
-    // trim the message to the leftover stderr width. It's a one-shot render
-    // (no `\r` redraw), but a wrapped completion line still reads ragged.
-    let avail = crate::utils::term::term_cols_stderr()
-        .saturating_sub(14)
-        .max(8);
-    let msg = console::truncate_str(&msg, avail, "…").into_owned();
-    // Set the prefix + message *before* swapping the style: the in-progress
-    // template ignores both slots, so this stays invisible until the style
-    // swap, which then renders the final line in one shot. Doing it the other
-    // way round flashes a `✓ [elapsed]` frame with an empty message first.
-    pb.set_prefix(prefix);
-    pb.set_message(msg);
-    pb.set_style(
-        ProgressStyle::default_bar()
-            .template("{prefix} [{elapsed_precise}] {msg}")
-            .expect("valid finish template"),
-    );
-    pb.finish();
-}
-
-/// Log WAF block statistics gathered during the scan (debug builds only).
-fn log_waf_block_stats(target: &Target) {
-    if crate::DEBUG.load(Ordering::Relaxed) {
-        let total_waf_blocks = crate::WAF_BLOCK_COUNT.load(Ordering::Relaxed);
-        if total_waf_blocks > 0 {
-            eprintln!(
-                "[*] WAF block stats: {} total blocks detected during scan of {}",
-                total_waf_blocks, target.url,
-            );
-        }
-    }
 }
 
 /// Collapse this target's R findings that are already proven by one of its
